merge feat into main

+++ src/api/v1/routes/answer.py
@@ -3,6 +3,7 @@ from uuid import UUID
 
 from fastapi import APIRouter, Depends, HTTPException, Path, status
 
+from src.core.logger import logger
 from src.db.session import DBSession
 from src.domain.schemas.answer import AnswerCreate, AnswerUpdate
 from src.services.answer import AnswerService
@@ -21,7 +22,6 @@ router = APIRouter(
 )
 async def create_answer(db: DBSession, data: AnswerCreate):
     _service = AnswerService(session=db)
-
     response = await _service.create_answer(data=data)
     if not response:
         raise HTTPException(
@@ -38,10 +38,11 @@ async def create_answer(db: DBSession, data: AnswerCreate):
 )
 async def get_answer(db: DBSession, answer_id: Annotated[UUID, Path()]):
     _service = AnswerService(session=db)
-
     response = await _service.get_answer(answer_id=answer_id)
     if not response:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="answer not found")
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="answer not found"
+        )
     else:
         return response
 
@@ -50,12 +51,23 @@ async def get_answer(db: DBSession, answer_id: Annotated[UUID, Path()]):
     "/{answer_id}/",
     status_code=status.HTTP_200_OK,
 )
-async def update_answer(db: DBSession, answer_id: Annotated[UUID, Path()], data: AnswerUpdate):
-    _service = AnswerService(session=db)
+async def update_answer(
+    db: DBSession,
+    answer_id: Annotated[UUID, Path()],
+    data: AnswerUpdate,
+    is_answer_owner: bool = Depends(AuthService.is_answer_owner),
+):
+    if not is_answer_owner:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail="Access denied"
+        )
 
+    _service = AnswerService(session=db)
     response = await _service.update_answer(answer_id=answer_id, data=data)
     if not response:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="answer not found")
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="answer not found"
+        )
     else:
         return response
 
@@ -64,11 +76,21 @@ async def update_answer(db: DBSession, answer_id: Annotated[UUID, Path()], data:
     "/{answer_id}/",
     status_code=status.HTTP_200_OK,
 )
-async def delete_answer(db: DBSession, answer_id: Annotated[UUID, Path()]):
-    _service = AnswerService(session=db)
+async def delete_answer(
+    db: DBSession,
+    answer_id: Annotated[UUID, Path()],
+    is_answer_owner: bool = Depends(AuthService.is_answer_owner),
+):
+    if not is_answer_owner:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail="Access denied"
+        )
 
+    _service = AnswerService(session=db)
     response = await _service.delete_answer(answer_id=answer_id)
     if not response:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Answer not found")
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Answer not found"
+        )
     else:
         return response
